pyproject.toml
[tool.poetry]
name = "stadium-booking"
version = "0.1.0"
description = ""
authors = ["Khasanjon <khasanjon.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
